requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-dotenv==1.0.1
openai==1.44.0
azure-search-documents==11.6.0
azure-core==1.30.2
pydantic==2.8.2
httpx==0.27.2
